requirements.txt
pypdf2
pdf2image
Pillow
pytesseract